refactored cd builtin
