+++ src/utils/memory.c
@@ -1,0 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   memory.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: spenning <spenning@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/06/24 18:43:01 by spenning          #+#    #+#             */
+/*   Updated: 2024/06/24 19:07:06 by spenning         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../include/minishell.h"
+
+void	free_char_array(char **arr)
+{
+	int	index;
+
+	index = 0;
+	while (arr[index] != NULL)
+	{
+		free(arr[index]);
+		arr[index] = 0;
+		index++;
+	}
+	free(arr);
+	arr = NULL;
+}
